Juggling files for current relevance.
